+++ Mods/SOD.LifeAndLiving/Patches/GameplayControlsPatches.cs
@@ -1,4 +1,5 @@
 ﻿using HarmonyLib;
+using System;
 
 namespace SOD.LifeAndLiving.Patches
 {
@@ -17,12 +18,23 @@ namespace SOD.LifeAndLiving.Patches
             {
                 var reduction = Plugin.Instance.Config.PayoutReductionMurders;
                 int count = 0;
+                int totalMin = 0, totalMax = 0;
                 foreach (var resolveQuestion in __instance.murderResolveQuestions)
                 {
-                    var min = resolveQuestion.rewardRange.x - (resolveQuestion.rewardRange.x / 100 * reduction);
-                    var max = resolveQuestion.rewardRange.y - (resolveQuestion.rewardRange.y / 100 * reduction);
+                    var min = (int)resolveQuestion.rewardRange.x - (int)(resolveQuestion.rewardRange.x / 100 * reduction);
+                    var max = (int)resolveQuestion.rewardRange.y - (int)(resolveQuestion.rewardRange.y / 100 * reduction);
                     resolveQuestion.rewardRange = new UnityEngine.Vector2(min, max);
+                    totalMin += min;
+                    totalMax += max;
                     count++;
+                }
+
+                // Fall-back to make murder reward atleast 50€
+                if (totalMin < 50 || totalMax < 50)
+                {
+                    var totalPerQuestion = (int)Math.Ceiling(50f / __instance.murderResolveQuestions.Count);
+                    foreach (var resolveQuestion in __instance.murderResolveQuestions)
+                        resolveQuestion.rewardRange = new UnityEngine.Vector2(totalPerQuestion, totalPerQuestion);
                 }
                 Plugin.Log.LogInfo($"Reduced \"{count}\" murder resolve question payouts.");
             }